moved N to more sensible place
