+++ go/lib/ctrl/seg/extensions/pqabeaconing/BUILD.bazel
@@ -2,7 +2,10 @@ load("//lint:go.bzl", "go_library", "go_test")
 
 go_library(
     name = "go_default_library",
-    srcs = ["pqa.go"],
+    srcs = [
+        "pb.go",
+        "pqa.go",
+    ],
     importpath = "github.com/scionproto/scion/go/lib/ctrl/seg/extensions/pqabeaconing",
     visibility = ["//visibility:public"],
     deps = ["//go/pkg/proto/control_plane:go_default_library"],
