+++ go/pkg/storage/pqa/memory/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//go/cs/ifstate:go_default_library",
         "//go/lib/addr:go_default_library",
         "//go/lib/common:go_default_library",
+        "//go/lib/ctrl/seg/extensions/pqabeaconing:go_default_library",
         "//go/lib/infra/modules/db:go_default_library",
         "//go/lib/log:go_default_library",
         "//go/lib/serrors:go_default_library",
